requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework
PyMySQL
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
